third_party/blink/web_tests/fast/parser/style-in-create-contextual-fragment-head-expected.txt

| <html>
|   <head>
|     <script>
|       src="../../resources/dump-as-markup.js"
|     "
"
|     <title>
|       "hi"
|     <style>
|     "XXX"
|     <div>
|       "ttt"
|   <body>
|     "
"
|     <script>
|       "
var range = document.createRange();
range.selectNode(document.head);
var documentFragment = range.createContextualFragment("<title>hi</title><style></style><body>XXX<div>ttt</div></body>");
document.head.appendChild(documentFragment);
"
|     "

"
